feat(experiments): probe matrix for 4 more adapters + final results (#164)
