+++ experiments/acp-probe/results/claude-agent-acp.log
@@ -1,0 +1,24 @@
+[claude-agent-acp] running c1_initialize...
+  → C1: pass — protocol_version=1; load_session=True
+[claude-agent-acp] running c2_new_session_cwd...
+  → C2: pass — session_id=5881affb-8a0a-40b3-8e11-f5da2825a302
+[claude-agent-acp] running c3_streaming_chunks...
+  → C3: pass — 2 agent_message_chunk frames, stop_reason=end_turn
+[claude-agent-acp] running c4_tool_call_lifecycle...
+  → C4: pass — 1 start, 3 progress
+[claude-agent-acp] running c5_request_permission...
+  → C5: partial — 0 request_permission round-trips (agent auto-allowed)
+[claude-agent-acp] running c6_cancel_midturn...
+  → C6: pass — stop_reason=cancelled
+[claude-agent-acp] running c7_load_session_replay...
+  → C7: pass — 2 replayed chunks, marker_found=True
+[claude-agent-acp] running c8_close_session...
+  → C8: pass — close_session returned cleanly
+[claude-agent-acp] running c9_fs_read_write...
+  → C9: pass — file_after='after'
+[claude-agent-acp] running c10_terminal...
+  → C10: n/a — terminal probe requires RecordingClient terminal handlers; deferred
+[claude-agent-acp] running c11_plan_updates...
+  → C11: fail — 0 plan updates
+[claude-agent-acp] running c12_content_blocks...
+  → C12: pass — image=True, audio=False
