+++ experiments/acp-probe/results/claude-agent-acp.md
@@ -1,0 +1,13 @@
+# claude-agent-acp probe results
+- C1: **pass** — protocol_version=1; load_session=True
+- C2: **pass** — session_id=5881affb-8a0a-40b3-8e11-f5da2825a302
+- C3: **pass** — 2 agent_message_chunk frames, stop_reason=end_turn
+- C4: **pass** — 1 start, 3 progress
+- C5: **partial** — 0 request_permission round-trips (agent auto-allowed)
+- C6: **pass** — stop_reason=cancelled
+- C7: **pass** — 2 replayed chunks, marker_found=True
+- C8: **pass** — close_session returned cleanly
+- C9: **pass** — file_after='after'
+- C10: **n/a** — terminal probe requires RecordingClient terminal handlers; deferred
+- C11: **fail** — 0 plan updates
+- C12: **pass** — image=True, audio=False
