+++ experiments/acp-probe/results/codex-acp.md
@@ -1,0 +1,13 @@
+# codex-acp probe results
+- C1: **pass** — protocol_version=1; load_session=True
+- C2: **pass** — session_id=019e37b1-6600-70f0-8fdd-adf2f579fef0
+- C3: **pass** — 9 agent_message_chunk frames, stop_reason=end_turn
+- C4: **pass** — 1 start, 1 progress
+- C5: **pass** — 1 request_permission round-trips (OK)
+- C6: **pass** — stop_reason=cancelled
+- C7: **pass** — 2 replayed chunks, marker_found=True
+- C8: **pass** — close_session returned cleanly
+- C9: **pass** — file_after='after\n'
+- C10: **n/a** — terminal probe requires RecordingClient terminal handlers; deferred
+- C11: **fail** — 0 plan updates
+- C12: **pass** — image=True, audio=False
