+++ experiments/acp-probe/results/copilot.log
@@ -1,0 +1,24 @@
+[copilot] running c1_initialize...
+  → C1: pass — protocol_version=1; load_session=True
+[copilot] running c2_new_session_cwd...
+  → C2: pass — session_id=4454db3a-2c2a-4e7b-9bb6-b9e48869e417
+[copilot] running c3_streaming_chunks...
+  → C3: pass — 9 agent_message_chunk frames, stop_reason=end_turn
+[copilot] running c4_tool_call_lifecycle...
+  → C4: pass — 1 start, 1 progress
+[copilot] running c5_request_permission...
+  → C5: pass — 1 request_permission round-trips (OK)
+[copilot] running c6_cancel_midturn...
+  → C6: partial — stop_reason=end_turn
+[copilot] running c7_load_session_replay...
+  → C7: pass — 2 replayed chunks, marker_found=True
+[copilot] running c8_close_session...
+  → C8: n/a — agent does not advertise session_capabilities.close
+[copilot] running c9_fs_read_write...
+  → C9: pass — file_after='after'
+[copilot] running c10_terminal...
+  → C10: n/a — terminal probe requires RecordingClient terminal handlers; deferred
+[copilot] running c11_plan_updates...
+  → C11: pass — 1 plan updates
+[copilot] running c12_content_blocks...
+  → C12: pass — image=True, audio=False
