+++ experiments/acp-probe/results/pi-acp.log
@@ -1,0 +1,36 @@
+[pi-acp] running c1_initialize...
+  → C1: pass — protocol_version=1; load_session=True
+[pi-acp] running c2_new_session_cwd...
+ERROR:root:Receive loop failed
+Traceback (most recent call last):
+  File "/Users/prass/development/projects/repowire/.venv/lib/python3.13/site-packages/acp/task/supervisor.py", line 51, in _on_done
+    task.result()
+    ~~~~~~~~~~~^^
+  File "/Users/prass/development/projects/repowire/.venv/lib/python3.13/site-packages/acp/connection.py", line 168, in _receive_loop
+    await self._process_message(message)
+  File "/Users/prass/development/projects/repowire/.venv/lib/python3.13/site-packages/acp/connection.py", line 182, in _process_message
+    await self._queue.publish(RpcTask(RpcTaskKind.NOTIFICATION, message))
+  File "/Users/prass/development/projects/repowire/.venv/lib/python3.13/site-packages/acp/task/queue.py", line 35, in publish
+    raise RuntimeError(msg)
+RuntimeError: mssage queue already closed
+  → C2: pass — session_id=019e37b3-8ff5-72cd-98e4-cea83925660d
+[pi-acp] running c3_streaming_chunks...
+  → C3: pass — 11 agent_message_chunk frames, stop_reason=end_turn
+[pi-acp] running c4_tool_call_lifecycle...
+  → C4: pass — 2 start, 53 progress
+[pi-acp] running c5_request_permission...
+  → C5: partial — 0 request_permission round-trips (agent auto-allowed)
+[pi-acp] running c6_cancel_midturn...
+  → C6: pass — stop_reason=cancelled
+[pi-acp] running c7_load_session_replay...
+  → C7: pass — 2 replayed chunks, marker_found=True
+[pi-acp] running c8_close_session...
+  → C8: n/a — agent does not advertise session_capabilities.close
+[pi-acp] running c9_fs_read_write...
+  → C9: pass — file_after='after'
+[pi-acp] running c10_terminal...
+  → C10: n/a — terminal probe requires RecordingClient terminal handlers; deferred
+[pi-acp] running c11_plan_updates...
+  → C11: fail — 0 plan updates
+[pi-acp] running c12_content_blocks...
+  → C12: pass — image=True, audio=False
