+++ experiments/acp-probe/results/pi-acp.md
@@ -1,0 +1,13 @@
+# pi-acp probe results
+- C1: **pass** — protocol_version=1; load_session=True
+- C2: **pass** — session_id=019e37b3-8ff5-72cd-98e4-cea83925660d
+- C3: **pass** — 11 agent_message_chunk frames, stop_reason=end_turn
+- C4: **pass** — 2 start, 53 progress
+- C5: **partial** — 0 request_permission round-trips (agent auto-allowed)
+- C6: **pass** — stop_reason=cancelled
+- C7: **pass** — 2 replayed chunks, marker_found=True
+- C8: **n/a** — agent does not advertise session_capabilities.close
+- C9: **pass** — file_after='after'
+- C10: **n/a** — terminal probe requires RecordingClient terminal handlers; deferred
+- C11: **fail** — 0 plan updates
+- C12: **pass** — image=True, audio=False
